biocypher/biocypher_cli/requirements.txt
click>=8.0.0
rich>=13.0.0
cryptography>=41.0.0
PyYAML>=6.0
tqdm>=4.65.0